single-responsability-principle/src/main/java/com/victorbarreto/after/Store.java
package com.victorbarreto.after;

import java.util.HashMap;
import java.util.Map;

//Stores data in memory
public class Store {

    private static final Map<String, User> STORAGE = new HashMap<>();

    public void store(User user) {
        if (user == null) {
            throw new IllegalArgumentException("User cannot be null");
        }
        synchronized(STORAGE) {
            STORAGE.put(user.getName(), user);
        }
    }

    public User getUser(String name) {
        synchronized(STORAGE) {
            return STORAGE.get(name);
        }
    }
}